Add files via upload
Seaborn is a library for making statistical graphics in Python. It is built on top of matplotlib and closely integrated with pandas data structures.

Here is some of the functionality that seaborn offers:

    A dataset-oriented API for examining relationships between multiple variables
    Specialized support for using categorical variables to show observations or aggregate statistics
    Options for visualizing univariate or bivariate distributions and for comparing them between subsets of data
    Automatic estimation and plotting of linear regression models for different kinds dependent variables
    Convenient views onto the overall structure of complex datasets
    High-level abstractions for structuring multi-plot grids that let you easily build complex visualizations
    Concise control over matplotlib figure styling with several built-in themes
    Tools for choosing color palettes that faithfully reveal patterns in your data
Seaborn aims to make visualization a central part of exploring and understanding data. Its dataset-oriented plotting functions operate on dataframes and arrays containing whole datasets and internally perform the necessary semantic mapping and statistical aggregation to produce informative plots.